MakeUpdate

+++ src/game_values.cpp
@@ -1,7 +1,7 @@
 #include "game_values.h"
 
 namespace game_values {
-int window = 0;	
+short window = 0;	
 FT_Library  library;
 };
